requirements.txt
fastapi
uvicorn
python-dotenv
PyMuPDF
langchain
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
google-generativeai
langchain-google-genai
python-multipart
